Fix deletion when no perms

+++ index.js
@@ -136,7 +136,10 @@ client.on('messageCreate', async (message) => {
 	}
 	// If the session is processing, don't do anything
 	if (sessions[message.channelId].processing) {
-		message.delete();
+		// if the bot has perms attempt to delete the message
+		if (message.guild.me.permissions.has("MANAGE_MESSAGES")) {
+			message.delete();
+		}
 		return message.author.send(lang.busy).then((msg) => {
 			setTimeout(() => {
 				msg.delete();
